+++ src/triggers/row_update.js
@@ -1,4 +1,7 @@
+const _CONST = require('../const')
 const ctx = require('../ctx')
+const {ResponseThrottleInfo} = require('../lib')
+
 const _ = require('lodash')
 
 /**
@@ -11,7 +14,11 @@ const _ = require('lodash')
  * @return {Promise<Array<{object}>|Array<{object}>|number|SQLResultSetRowList|HTMLCollectionOf<HTMLTableRowElement>|string>}
  */
 const perform = async (z, bundle) => {
+
   const dtableCtx = await ctx.acquireDtableAppAccess(z, bundle)
+
+  const logTag = `[${bundle.__zTS}] triggers.row_update`;
+  z.console.time(logTag);
 
   /** @type {ZapierZRequestResponse} */
   const response = await z.request({
@@ -21,10 +28,15 @@ const perform = async (z, bundle) => {
   })
 
   let rows = response.data.rows
-  rows = _.orderBy(rows, ['_mtime'], ['desc'])
 
   const meta = bundle.meta
 
+  z.console.timeLog(logTag, `rows(${new ResponseThrottleInfo(response)}) length=${rows.length} meta: limit=${meta && meta.limit} isLoadingSample=${meta && meta.isLoadingSample}`)
+  if (0 === rows.length) {
+    return rows;
+  }
+
+  rows = _.orderBy(rows, ['_mtime'], ['desc'])
   if (meta && meta.isLoadingSample) {
     rows.splice(meta.limit || 3)
   }
@@ -36,8 +48,31 @@ const perform = async (z, bundle) => {
     return o
   })
 
+  const unfilteredLength = rows.length;
+
+  const featureMTime = _CONST.FEATURE[_CONST.FEATURE_MTIME_FILTER] || undefined;
+  if (featureMTime && featureMTime.enabled) {
+    if (bundle._testFeature && bundle._testFeature[_CONST.FEATURE_MTIME_FILTER]) {
+      if (bundle._testFeature[_CONST.FEATURE_MTIME_FILTER].captureRowsBeforeFilter) {
+        bundle._testFeature[_CONST.FEATURE_MTIME_FILTER].capturedRows = rows;
+      } else {
+        bundle._testFeature[_CONST.FEATURE_MTIME_FILTER].capturedRows = null;
+      }
+    }
+    const mTimeFilterMinutes = featureMTime.minutes
+    const MILLISECONDS_PER_MINUTE = 60000;
+    const floorEpochMilliseconds = (new Date).valueOf() - (mTimeFilterMinutes * MILLISECONDS_PER_MINUTE);
+
+    rows = _.filter(rows, (o) => {
+      return Date.parse(o.row_mtime) >= floorEpochMilliseconds;
+    })
+    z.console.timeLog(logTag, `filtered rows length: ${rows.length} (offset=${unfilteredLength - rows.length} minutes=${mTimeFilterMinutes})`)
+  }
+
   rows = await ctx.acquireFileNoAuthLinks(z, bundle, tableMetadata.columns, rows)
   rows = await ctx.acquireLinkColumnsData(z, bundle, tableMetadata.columns, rows)
+
+  z.console.timeLog(logTag, `rows length: ${rows && rows.length}`)
 
   return rows
 }